Timestamp parsing in clean() should be done based on appendTimestamp flag.
But, what if appendTimeStamp is true but there is no timestamp appended to JSESSION?
