+++ session-replacement/src/main/java/com/amadeus/session/servlet/BaseSessionTracking.java
@@ -1,7 +1,5 @@
 package com.amadeus.session.servlet;
 
-import java.nio.ByteBuffer;
-
 import javax.servlet.ServletContext;
 
 import com.amadeus.session.RandomIdProvider;
@@ -10,7 +8,6 @@ import com.amadeus.session.SessionConfiguration;
 import com.amadeus.session.SessionIdProvider;
 import com.amadeus.session.SessionTracking;
 import com.amadeus.session.UuidProvider;
-import com.amadeus.session.Base64MaskingHelper;
 
 /**
  * This base class for session ID tracking that can be configured via
@@ -47,11 +44,8 @@ public abstract class BaseSessionTracking implements SessionTracking {
   public String newId() {
     String newId = idProvider.newId();
     if (appendTimestamp) {
-      StringBuilder suffixedId = new StringBuilder(newId);
-      ByteBuffer buffer = ByteBuffer.allocate(Long.BYTES);
-      newId = suffixedId.append(SESSION_ID_TIMESTAMP_SEPARATOR)
-              .append(Base64MaskingHelper.encode(buffer.putLong(System.currentTimeMillis()).array()))
-              .toString();
+        StringBuilder suffixedId = new StringBuilder(newId.length() + 11).append(newId);
+        newId = suffixedId.append(SESSION_ID_TIMESTAMP_SEPARATOR).append(System.currentTimeMillis()).toString();
     }
     return newId;
   }
@@ -73,7 +67,7 @@ public abstract class BaseSessionTracking implements SessionTracking {
     String timeStamp = "";
     String cleanValue = value;
     int separatorIndex = value.lastIndexOf(SESSION_ID_TIMESTAMP_SEPARATOR);
-    if (separatorIndex != -1) {
+    if (appendTimestamp && separatorIndex != -1) {
         timeStamp = value.substring(separatorIndex);
         cleanValue = value.substring(0, separatorIndex);
     }
