+++ session-replacement/src/test/java/com/amadeus/session/servlet/TestUrlSessionTracking.java
@@ -62,10 +62,24 @@ public class TestUrlSessionTracking {
     when(hsr.getPathInfo()).thenReturn(";somesession="+uuid);
     String id = urlSessionTracking.retrieveId(request);
     assertEquals(uuid.toString(), id);
+
     String sessionIdWithTimestamp = uuid.toString() + BaseSessionTracking.SESSION_ID_TIMESTAMP_SEPARATOR + System.currentTimeMillis();
     String invalidSessionIdWithTimestamp = uuid.toString() + "-abcdefgh" + BaseSessionTracking.SESSION_ID_TIMESTAMP_SEPARATOR + System.currentTimeMillis();
+    String sessionIdWithoutTimestamp = uuid.toString();
+
+    when(hsr.getPathInfo()).thenReturn(";somesession="+sessionIdWithTimestamp);
+    assertNull(urlSessionTracking.retrieveId(request));
+    when(hsr.getPathInfo()).thenReturn(";somesession="+sessionIdWithoutTimestamp);
+    assertEquals(sessionIdWithoutTimestamp, urlSessionTracking.retrieveId(request));
+    when(hsr.getPathInfo()).thenReturn(";somesession="+invalidSessionIdWithTimestamp);
+    assertNull(urlSessionTracking.retrieveId(request));
+
+    sc.setTimestampSufix(true);
+    urlSessionTracking.configure(sc);
     when(hsr.getPathInfo()).thenReturn(";somesession="+sessionIdWithTimestamp);
     assertEquals(sessionIdWithTimestamp, urlSessionTracking.retrieveId(request));
+    when(hsr.getPathInfo()).thenReturn(";somesession="+sessionIdWithoutTimestamp);
+    assertEquals(sessionIdWithoutTimestamp, urlSessionTracking.retrieveId(request));
     when(hsr.getPathInfo()).thenReturn(";somesession="+invalidSessionIdWithTimestamp);
     assertNull(urlSessionTracking.retrieveId(request));
   }
